Designed and fleshed out the dashboard/home page. It contains links to other pages.
I assume we are planning on removing the top navigation bar and having our app only use the left sidebar and navigation from this dashboard. I figured this would be a good landing page for people. Let me know your thoughts on this/the navigational design of the website as a whole if you have any other ideas.
Oh I also added the doctors page to the top nav bar for consistency even though it will eventually be removed later.